WellKnown: Update Matrix server address
This patch changes the Matrix federation server and client api
address from matrix.serenityos.org:8443 to
serenityos.matrix.nwex.de:443.

Using the non standard port isn't ideal in the infrastructure
setup, as such streamlining the subdomain namespace with other
Matrix servers is an obvious choice. The Element Web deployment
on matrix.serenityos.org:443 will stay independent of this. To
prevent XSS it wouldn't be possible to also run federation on
the same origin.